Adding a StartScreen component Creating a TabBar functionality
