+++ modules/FoodBagService.js
@@ -1,8 +1,8 @@
 import axios from 'axios'
 import store from '../state/store/store'
 
-const API_URL = 'https://git.heroku.com/weca.git'
-// const API_URL = 'http://localhost:3000/api'
+// const API_URL = 'https://git.heroku.com/weca.git'
+ const API_URL = 'http://localhost:3000/api'
 const FoodBagService = {
   async index() {
     let response = await axios.get(API_URL + '/foodbags')
